% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{add_group}
\alias{add_group}
\title{Group data}
\usage{
add_group(dat, group_names)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}}
function.}

\item{group_names}{a single character string or character vector
specifying the names of the columns to group the data by.}
}
\value{
\code{\link{raw_data}} object.
}
\description{
\code{add_group()} groups data by one or more columns.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")

dat <- add_group(dat, c("group", "species"))

}
\seealso{
\code{\link[=read_data]{read_data()}}
}
