% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{build_models}
\alias{build_models}
\title{Build models}
\usage{
build_models(dat, response, level = NULL)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}}
function.}

\item{response}{a string specifying the response variable.}

\item{level}{a string specifying which level of the response variable to
consider. Required only if response variable has over two levels.}
}
\description{
\code{build_models()} returns the linear models with and without SLOPE coś tam
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
build_models(dat, "group", "2")

}
