% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{calcuate_CV}
\alias{calcuate_CV}
\title{Calculate CV for different QC samples depending on metabolite}
\usage{
calcuate_CV(dat)
}
\arguments{
\item{dat}{A raw_data object.}
}
\description{
Calculate CV for different QC samples depending on metabolite
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat[,18:26] <- apply(data[,18:26], 2, function(dat){
dat[which(dat == "< LOD" | dat == "> ULOQ" | is.na(dat) | dat == "< LLOQ" | dat == "NA")] <- 0
dat
})
calcuate_CV(dat)

}
