% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_LODs.R
\name{complete_data}
\alias{complete_data}
\title{Complete not quantified values}
\usage{
complete_data(
  dat,
  LOD_method = NULL,
  LLOQ_method = NULL,
  ULOQ_method = NULL,
  LOD_type = "calc"
)
}
\arguments{
\item{dat}{}

\item{LOD_method}{description}

\item{LLOQ_method}{description}

\item{ULOQ_method}{description}

\item{LOD_type}{}
}
\description{
This function completes missing values related to limits of
quantification or detection.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat)

}
