% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_LODs.R
\name{complete_LLOQ}
\alias{complete_LLOQ}
\title{Complete values below limit of quantification}
\usage{
complete_LLOQ(gathered_data, LLOQ_method, LOD_vals)
}
\arguments{
\item{gathered_data}{a long-format data table.}

\item{LLOQ_method}{a character string specifying the imputation method to be
applied to \verb{< LLOQ} values, or \code{NULL} if these values should not be imputed.
Currently, the only available method is: \code{limit}.}

\item{LOD_vals}{a long-format table containing limits of
detection/quantification and corresponding plate bar codes.}
}
\description{
\code{complete_LLOQ()} imputes values below limit of quantification
("< LLOQ").
}
\keyword{internal}
