% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_LODs.R
\name{complete_ULOQ}
\alias{complete_ULOQ}
\title{Complete values above limit of quantification}
\usage{
complete_ULOQ(gathered_data, ULOQ_method, LOD_vals)
}
\arguments{
\item{gathered_data}{a long-format data table.}

\item{ULOQ_method}{a character string specifying the imputation method to be
applied to \verb{> ULOQ} values, or \code{NULL} if these values should not be imputed.
Available methods are: \code{limit}, \verb{third quartile} and \verb{scaled random}.}

\item{LOD_vals}{a long-format table containing limits of
detection/quantification and corresponding plate bar codes.}
}
\description{
\code{complete_ULOQ()} imputes values below limit of quantification
("> ULOQ").
}
\keyword{internal}
