% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_beeswarm_plot}
\alias{create_beeswarm_plot}
\title{Beeswarm Plot of Metabolite Values}
\usage{
create_beeswarm_plot(dat, metabolite)
}
\arguments{
\item{dat}{A \code{raw_data} object, the output of the \code{\link[=read_data]{read_data()}} function.
The data should contain metabolite values and sample information.}

\item{metabolite}{A character string specifying the name of the metabolite for which the beeswarm plot should be created.}
}
\description{
This function creates a beeswarm plot for a specified metabolite, allowing visualization
of the distribution of metabolite values across different sample types or groups.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
path <- get_example_data("dwa_sety_example.xlsx")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_beeswarm_plot(dat, "C0")

}
