% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Boxplot of individual metabolite values}
\usage{
create_boxplot(dat, metabolite)
}
\arguments{
\item{metabolite}{The name of metabolite of interest.}
}
\description{
Boxplot of individual metabolite values
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_boxplot(attr(dat, "completed"), "C0")

}
