% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_boxplots}
\alias{create_boxplots}
\title{Boxplots of individual metabolites values}
\usage{
create_boxplots(dat)
}
\description{
Boxplots of individual metabolites values
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_boxplots(dat)

}
