% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_correlations_heatmap}
\alias{create_correlations_heatmap}
\title{Heatmap of correlations between metabolites}
\usage{
create_correlations_heatmap(dat)
}
\description{
Heatmap of correlations between metabolites
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat[,18:26] <- sapply(dat[,18:26], function(col){
ifelse(is.na(as.numeric(col)), 0, as.numeric(col))
})
create_correlations_heatmap(dat)

}
