% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_density_with_lod}
\alias{create_density_with_lod}
\title{This function creates a density plot for a specified metabolite, overlaying a
vertical dashed line indicating the Limit of Detection (LOD) cutoff. The LOD
value is derived from the \code{LOD_table} attribute of the \code{dat} object.}
\usage{
create_density_with_lod(dat, metabolite_name)
}
\arguments{
\item{dat}{A \code{raw_data} object, the output of the \code{\link[=read_data]{read_data()}} function. The
data should contain the metabolite values and LOD information.}

\item{metabolite_name}{A character string specifying the name of the
metabolite for which the histogram should be created.}
}
\description{
This function creates a density plot for a specified metabolite, overlaying a
vertical dashed line indicating the Limit of Detection (LOD) cutoff. The LOD
value is derived from the \code{LOD_table} attribute of the \code{dat} object.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_density_with_lod(dat, "C0")

}
