% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_histogram}
\alias{create_histogram}
\title{Histogram of individual metabolite values before and after imputation}
\usage{
create_histogram(uncomp_metabo_vals, comp_metabo_vals, metabolite, bins, type)
}
\description{
Histogram of individual metabolite values before and after imputation
}
\keyword{internal}
