% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_histogram_with_lod}
\alias{create_histogram_with_lod}
\title{Estimated density of metabolite with LOD cut-off}
\usage{
create_histogram_with_lod(dat, metabolite_name)
}
\description{
Estimated density of metabolite with LOD cut-off
}
