% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_joint_NA_plot}
\alias{create_joint_NA_plot}
\title{ggplot object for plot_NA_percent with type "joint"}
\usage{
create_joint_NA_plot(dat)
}
\description{
ggplot object for plot_NA_percent with type "joint"
}
\keyword{internal}
