% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_long_metabolites_tibble}
\alias{create_long_metabolites_tibble}
\title{Long tibble with metabolites}
\usage{
create_long_metabolites_tibble(dat)
}
\description{
Long tibble with metabolites
}
\keyword{internal}
