% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_PCA_plot}
\alias{create_PCA_plot}
\title{PCA plot}
\usage{
create_PCA_plot(dat)
}
\description{
PCA plot
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_PCA_plot(attr(dat, "completed"))

}
