% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_plot_of_2_metabolites}
\alias{create_plot_of_2_metabolites}
\title{Plot of two metabolites}
\usage{
create_plot_of_2_metabolites(dat, metabolite1, metabolite2, interactive = TRUE)
}
\arguments{
\item{dat}{A \code{raw_data} object, the output of the \code{\link[=read_data]{read_data()}} function.
The data should be completed and filtered to include only samples of type
"Sample".}

\item{metabolite1}{words}

\item{metabolite2}{words}

\item{interactive}{logical indicating whether the plot should be interactive.
If TRUE, the plot includes interactive tooltips.}
}
\description{
Plot of two metabolites
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
print(create_plot_of_2_metabolites(dat, "C0", "C2"))

}
