% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_qqplots}
\alias{create_qqplots}
\title{Qqplots of individual metabolites values}
\usage{
create_qqplots(dat)
}
\description{
Qqplots of individual metabolites values
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
create_qqplots(dat)

}
