% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_venn_diagram}
\alias{create_venn_diagram}
\title{Venn diagram for group levels}
\usage{
create_venn_diagram(dat, threshold)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object, the output of \code{\link[=read_data]{read_data()}},
with a group specified using the \code{\link[=add_group]{add_group()}} function.}

\item{threshold}{a minimum ratio of metabolite missing values in one group
level for metabolite to be included in the diagram, given as decimal.}
}
\description{
This function creates Venn diagram, showing counts of metabolites having
ratios of missing values larger than the given threshold for each group level.
Function works only when group has up to 4 levels.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")
create_venn_diagram(dat, 0.1)

}
