% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_LODs.R
\name{general_min}
\alias{general_min}
\title{Calculate minimum ignoring character values}
\usage{
general_min(x)
}
\arguments{
\item{x}{a vector of observations.}
}
\description{
\code{general_min()} calculates minimum without values such as NA's,
values below or above LOD limit and so on.
}
\examples{
x <- c("<LOD", 5, 6, NA, 9, 16)
MetaboCrates:::general_min(x)

}
\keyword{internal}
