% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_LODs.R
\name{general_third_quartile}
\alias{general_third_quartile}
\title{Calculate third quartile ignoring character values}
\usage{
general_third_quartile(x)
}
\arguments{
\item{x}{a vector of observations}
}
\description{
This function calculates third quartile without values such as
NA's, values below or above LOD limit and so on.
}
\examples{
x <- c("<LOD", 5, 6, NA, 9, 16)
general_third_quartile(x)

}
\keyword{internal}
