% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{get_CV_to_remove}
\alias{get_CV_to_remove}
\title{Get CV to remove}
\usage{
get_CV_to_remove(dat, threshold)
}
\arguments{
\item{dat}{object with CV attribute}

\item{threshold}{words}
}
\description{
Returns metabolite names having more CV value
than the given threshold.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
dat <- calculate_CV(dat)
get_CV_to_remove(dat, 0.3)

}
