% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{Get path to example data}
\usage{
get_example_data(name)
}
\arguments{
\item{name}{Name of the file.}
}
\value{
Path to example data.
}
\description{
Allows to access example data.
}
\examples{
get_example_data("small_biocrates_example.xlsx")

}
