% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{get_info}
\alias{get_info}
\title{Get information about data}
\usage{
get_info(dat)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}}
function.}
}
\description{
\code{get_info()} returns information about the number of sample types,
missing values types and grouping (if applied).
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
cat(get_info(dat))

dat <- add_group(dat, "group")
cat(get_info(dat))

}
\seealso{
\code{\link[=read_data]{read_data()}}
}
