% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_modeling_data}
\alias{get_modeling_data}
\title{Get data cleaned for modeling}
\usage{
get_modeling_data(dat, response, level = NULL)
}
\description{
Get data cleaned for modeling
}
\keyword{internal}
