% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_models_info}
\alias{get_models_info}
\title{Get models summary}
\usage{
get_models_info(models)
}
\arguments{
\item{models}{the object returned by \code{\link[=build_models]{build_models()}}.}
}
\description{
\code{get_models_info()} returns the summaries for models returned by
\code{\link[=build_models]{build_models()}}.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
models <- build_models(dat, "group", "2")
get_models_info(models)

}
