% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_LODs.R
\name{match_plate_codes}
\alias{match_plate_codes}
\title{Find matching plate bar codes in the LOD table}
\usage{
match_plate_codes(LOD_table, sets)
}
\arguments{
\item{LOD_table}{an \code{LOD_table} attribute from \code{\link{raw_data}} object.}

\item{sets}{an unique character vector of plate bar codes (for example
'1036372116-1 | 1036372121-1').}
}
\value{
A long-format table containing limits of detection/quantification
and corresponding plate bar codes.
}
\description{
Find matching plate bar codes in the LOD table
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
sets <- unique(dat[["plate bar code"]])
MetaboCrates:::match_plate_codes(attr(dat, "LOD_table"), sets)

}
\keyword{internal}
