% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboCrates-package.R
\docType{package}
\name{MetaboCrates}
\alias{-package}
\alias{MetaboCrates}
\title{MetaboCrates}
\description{
The \code{MetaboCrates} package is a toolbox for processing of
Biocrates® metabolomics data
}
\author{
Krystyna Grzesiak, Joanna Pokora, Filip Gaj, Michal Burdukiewicz.
}
