% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboCrates_gui.R
\name{MetaboCrates_gui}
\alias{MetaboCrates_gui}
\title{MetaboCrates Graphical User Interface}
\usage{
MetaboCrates_gui()
}
\description{
Launches graphical user interface .
}
\section{Warning}{
 Any ad-blocking software may cause malfunctions.
}

\examples{
if(interactive()) {
  MetaboCrates_gui()
}

}
