% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_class.R
\name{new_raw_data}
\alias{new_raw_data}
\title{Creation of validated raw_data class}
\usage{
new_raw_data(
  metabolomics_matrix,
  LOD_table,
  NA_info,
  metabolites,
  samples,
  group,
  removed
)
}
\value{
\code{\link{raw_data}} object.
}
\description{
The wrapper function for the constructor of the
\code{\link{raw_data}} and its validator. Used in \code{\link{raw_data}}
function.
}
\keyword{internal}
