% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{pca_variance}
\alias{pca_variance}
\title{Plot of variance explained by principal components}
\usage{
pca_variance(
  dat,
  threshold,
  type = "sample_type",
  max_num = NULL,
  cumulative = TRUE
)
}
\arguments{
\item{dat}{a \code{raw_data} object, the output of the \code{\link[=read_data]{read_data()}} function.
The data have to be completed, for example using the \code{\link[=complete_data]{complete_data()}}
function.}

\item{threshold}{a decimal indicating the maximum cumulative variance to
include in the plot.}

\item{type}{a character specifying which rows to consider. The default is
\code{"sample_type"}, which uses all rows. When set to \code{"group"}, only
observations of type "sample" are included.}

\item{max_num}{an optional integer indicating the maximum number
of components to display.}

\item{cumulative}{logical. If \code{TRUE}, a line representing the cumulative
variance is shown on the plot.}
}
\description{
\code{pca_variance()} creates a barplot showing the variance explained by each
principal component in a Principal Component Analysis (PCA). It includes all
components for which the cumulative variance explained is below the
specified threshold, as well as the first component for which the cumulative
variance exceeds the threshold (or up to the specified maximum number, if
given). Optionally, a line representing the cumulative variance explained
can be included.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
pca_variance(dat, 0.8, max_num = 5)

}
