% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of missing metabolites values}
\usage{
plot_heatmap(dat)
}
\description{
Heatmap of missing metabolites values
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
plot_heatmap(dat)

}
