% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_mv_types}
\alias{plot_mv_types}
\title{Barplot of metabolomics numbers}
\usage{
plot_mv_types(dat)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}} function}
}
\description{
Barplot of metabolomics numbers
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
test_dat <- read_data(path)
plot_mv_types(test_dat)

}
