% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_class.R
\name{raw_data}
\alias{raw_data}
\title{raw_data class}
\usage{
raw_data(metabolomics_matrix, LOD_table, metabolites)
}
\arguments{
\item{metabolomics_matrix}{\code{\link{data.frame}}, matrix containing
biocrates data}

\item{LOD_table}{\code{\link{data.frame}}, the LOD table containing limits of
detection / quantification}

\item{metabolites}{\code{\link{character}}, vector of metabolites names
contained in the data.}
}
\value{
\code{\link{raw_data}} object.
}
\description{
raw_data class
}
\keyword{internal}
