% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read Biocrates data}
\usage{
read_data(path)
}
\arguments{
\item{path}{Path to the file containing Biocrates data imported from MetIDQ
or WebIDQ TODO: description.}
}
\value{
\code{\link{raw_data}} object.
}
\description{
Imports Biocrates data from the Excel file.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
read_data(path)

}
