% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{remove_LOD}
\alias{remove_LOD}
\title{Print Biocrates data}
\usage{
remove_LOD(raw_data, metabolites_to_remove)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}} function.}
}
\description{
Print Biocrates data
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
print(dat)

dat <- add_group(dat, "group")
print(dat)

}
\seealso{
\code{\link[=read_data]{read_data()}}
}
