% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{scale_color_metabocrates_continuous}
\alias{scale_color_metabocrates_continuous}
\title{Metabocrates scale color continuous}
\usage{
scale_color_metabocrates_continuous()
}
\description{
Metabocrates scale color continuous
}
\keyword{internal}
