% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{scale_fill_metabocrates_discrete}
\alias{scale_fill_metabocrates_discrete}
\title{MetaboCrates scale fill discrete}
\usage{
scale_fill_metabocrates_discrete(n = 12, ...)
}
\description{
MetaboCrates scale fill discrete
}
\keyword{internal}
