% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{show_ratios}
\alias{show_ratios}
\title{Show missing values ratios without those in the removed metabolites}
\usage{
show_ratios(dat)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}}
function.}
}
\description{
\code{show_ratios()} returns LOD ratios without metabolites in removed attribute.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- remove_metabolites(dat, "C0", type = "LOD")
show_ratios(dat)

}
