% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{split_model_dat}
\alias{split_model_dat}
\title{Split data to train and test}
\usage{
split_model_dat(modeling_dat, train_prop, get_train_obs_num = FALSE)
}
\description{
Split data to train and test
}
\keyword{internal}
