% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{unremove_all}
\alias{unremove_all}
\title{Setting the specified type in attribute \code{removed} to \code{NULL}}
\usage{
unremove_all(dat, type)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}}
function.}

\item{type}{a character string indicating the type from which to unremove
all metabolites.}
}
\description{
\code{unremove_all()} set all the metabolites from the specified type as not
removed.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
test_dat <- read_data(path)
test_dat <- remove_metabolites(test_dat, "C0", "LOD")
attr(test_dat, "removed")
attr(unremove_all(test_dat, "LOD"), "removed")

}
