% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{unremove_LOD}
\alias{unremove_LOD}
\title{Setting attribute removed to NULL}
\usage{
unremove_LOD(raw_data)
}
\arguments{
\item{raw_data}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}} function.}
}
\description{
Setting attribute removed to NULL
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
test_dat <- read_data(path)
attr(remove_LOD(test_dat), "removed")

}
