% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{unremove_metabolites}
\alias{unremove_metabolites}
\title{Setting attribute removed to NULL}
\usage{
unremove_metabolites(raw_data, type)
}
\arguments{
\item{raw_data}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}}
function.}

\item{type}{type of metabolites to remove}
}
\description{
Setting attribute removed to NULL
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
test_dat <- read_data(path)
attr(remove_metabolites(test_dat, "C0", "LOD"), "removed")
attr(unremove_metabolites(test_dat, "LOD"), "removed")

}
