% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_class.R
\name{validate_raw_data}
\alias{validate_raw_data}
\title{Validator on raw_data class}
\usage{
validate_raw_data(raw_data)
}
\arguments{
\item{raw_data}{\code{\link{raw_data}} object.}
}
\description{
Validator on the content of an raw_data object.
}
\keyword{internal}
