% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{add_group}
\alias{add_group}
\title{Specify group in the Biocrates data}
\usage{
add_group(dat, group_name)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}} function.}

\item{group_name}{a character name of a column from the data containing group.}
}
\value{
\code{\link{raw_data}} object.
}
\description{
Specify group in the Biocrates data
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")

}
\seealso{
\code{\link[=read_data]{read_data()}}
}
