% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{build_model}
\alias{build_model}
\title{Build linear models}
\usage{
build_model(
  dat,
  response,
  level = NULL,
  model = "SLOPE",
  nfolds = 5,
  train_prop = 0.6
)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}}
function.}

\item{response}{a string specifying the name of the response variable.}

\item{level}{a string specifying the name of the level to model when
the response has more than two levels.}
}
\description{
\code{build_models()} fits full and penalized logistic models using the
specified grouping column as the outcome.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
build_model(dat, "group", "2")

}
