% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{calculate_CV}
\alias{calculate_CV}
\title{Calculate CV for different QC samples depending on metabolite}
\usage{
calculate_CV(dat)
}
\arguments{
\item{dat}{A raw_data object.}
}
\description{
Calculate CV for different QC samples depending on metabolite
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
calculate_CV(dat)

}
