% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{check_values}
\alias{check_values}
\title{Convert weird values into NA's}
\usage{
check_values(
  metabolite_vals,
  special_signs = c("< LOD", "< LLOQ", "> ULOQ", "NA", "∞")
)
}
\arguments{
\item{metabolite_vals}{a \code{\link{character}} vector containing measured
metabolomite values.}

\item{special_signs}{a character vector of permitted special signs that should
not be converted into NA's. Default to "< LOD","< LLOQ", "> ULOQ", "NA" and "∞".}
}
\description{
Convert weird values into NA's
}
