% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_LODs.R
\name{complete_LLOQ}
\alias{complete_LLOQ}
\title{Complete values below limit of quantification}
\usage{
complete_LLOQ(gathered_data, method, LOD_vals)
}
\arguments{
\item{gathered_data}{description}

\item{method}{a character. Imputation method, one of "limit" or NULL meaning
changing missing values into NAs.}

\item{LOD_vals}{description}
}
\description{
This function imputes values below limit of quantification
("< LLOQ")
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)

}
\keyword{internal}
