% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_LODs.R
\name{complete_LOD}
\alias{complete_LOD}
\title{Complete values below limit of detection}
\usage{
complete_LOD(gathered_data, LOD_type, method, LOD_vals)
}
\arguments{
\item{gathered_data}{description}

\item{LOD_type}{a character. Type of LOD values form table ('OP' or 'calc').
It can be NULL depending on the \code{method} parameter.}

\item{method}{a character. Imputation method, one of "halfmin",  "random",
"halflimit", "limit" or NULL meaning changing missing values into NAs.}

\item{LOD_vals}{description}
}
\description{
Complete values below limit of detection
}
\keyword{internal}
