% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_correlations_heatmap}
\alias{create_correlations_heatmap}
\title{Heatmap of correlations between metabolites}
\usage{
create_correlations_heatmap(
  dat,
  threshold = 0.3,
  metabolites_to_display = "all",
  width_svg = 6,
  height_svg = 5,
  interactive = TRUE
)
}
\arguments{
\item{threshold}{A numeric value specifying the minimum absolute correlation
to display (only metabolites specified in metabolites_to_display are taken
into account).}

\item{metabolites_to_display}{A vector of names or number of metabolites to
display. If a number is provided, the first metabolites are selected.
Defaults to "all".}

\item{width_svg}{Width of plot in inches.}

\item{height_svg}{Height of plot in inches.}

\item{interactive}{If TRUE, the plot includes interactive tooltips.}
}
\description{
Heatmap of correlations between metabolites
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
print(create_correlations_heatmap(dat))

}
