% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_correlations_heatmap}
\alias{create_correlations_heatmap}
\title{Heatmap of correlations between metabolites}
\usage{
create_correlations_heatmap(
  dat,
  type = "completed",
  threshold = 0.3,
  metabolites_to_display = "all",
  interactive = TRUE
)
}
\arguments{
\item{dat}{a \code{raw_data} object, the output of the \code{\link[=read_data]{read_data()}} function.
The data have to be completed, for example using the \code{\link[=complete_data]{complete_data()}}
function.}

\item{type}{default to \code{completed}, which creates a heatmap of correlations
between the metabolites after imputation. If \code{both}, then correlations
between the metabolites before and after imputation are shown.}

\item{threshold}{a numeric value specifying the minimum absolute correlation
to display (only metabolites specified in metabolites_to_display are taken
into account).}

\item{metabolites_to_display}{a vector of names or number of metabolites to
display. If a number is provided, the first metabolites are selected.
Defaults to \code{all}.}
}
\description{
Heatmap of correlations between metabolites
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
print(create_correlations_heatmap(dat))
print(create_correlations_heatmap(dat, type = "both"))

}
