% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_correlations_heatmap}
\alias{create_correlations_heatmap}
\title{Heatmap of correlations between metabolites}
\usage{
create_correlations_heatmap(dat)
}
\description{
Heatmap of correlations between metabolites
}
