% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_distribution_plot}
\alias{create_distribution_plot}
\title{Histograms, density or beeswarm plots of individual metabolite values before and after imputation}
\usage{
create_distribution_plot(
  dat,
  metabolite,
  type = "histogram",
  bins = 30,
  histogram_type = "all",
  interactive = TRUE
)
}
\arguments{
\item{dat}{a \code{raw_data} object, the output of the \code{\link[=read_data]{read_data()}} function.
The data have to be completed, for example using the \code{\link[=complete_data]{complete_data()}}
function.}

\item{metabolite}{a name of the metabolite of interest.}

\item{type}{a type of the plot. Can be either "histogram" (default),
"density", or "beeswarm".}

\item{bins}{the number of bins for the histogram plot; 30 if not specified.}

\item{histogram_type}{if "all' (default), the histogram displays all values
after imputation. If "imputed", it shows only the values that were imputed.}

\item{interactive}{logical. If \code{TRUE} (default), a ggiraph interactive plot
is returned; otherwise, a standard ggplot object is produced.}
}
\description{
\code{create_distribution_plot()} creates density plots of metabolite values
before and after imputation (with the sample limit of detection), beeswarm
plots or histograms of the values before and after imputation. Histogram
can also show only imputed values.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
print(create_distribution_plot(dat, "C3"))
print(create_distribution_plot(dat, "C3", type = "density"))
print(create_distribution_plot(dat, "C3", type = "beeswarm"))

}
