% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_distribution_plot}
\alias{create_distribution_plot}
\title{Histograms or density plots of individual metabolite values before and after imputation}
\usage{
create_distribution_plot(dat, metabolite, type = "histogram", bins = 30)
}
\arguments{
\item{metabolite}{A name of metabolite of interest.}

\item{type}{A type of the plot. Can be "histogram" (default) or "density".}

\item{bins}{The number of bins for the histogram plot, 30 if not specified.}
}
\description{
This function creates histograms of metabolite values before and after
imputation or density plots with the sample limit of detection.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_distribution_plot(dat, "C0")
create_distribution_plot(dat, "C0", type = "density")

}
