% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_histogram}
\alias{create_histogram}
\title{Histogram of individual metabolite values}
\usage{
create_histogram(dat, metabolite, bins_num = 30)
}
\arguments{
\item{metabolite}{The name of metabolite of interest.}

\item{bins_num}{Number of bins on a histogram. Defaults to 30.}
}
\description{
Histogram of individual metabolite values
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_histogram(dat, "C0")

}
