% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_PCA_plot}
\alias{create_PCA_plot}
\title{PCA plot}
\usage{
create_PCA_plot(
  dat,
  type = "scatterplot",
  group_by = "sample_type",
  types_to_display = "all",
  threshold = 0,
  interactive = TRUE
)
}
\arguments{
\item{dat}{a \code{raw_data} object, the output of the \code{\link[=read_data]{read_data()}} function.
The data have to be completed, for example using the \code{\link[=complete_data]{complete_data()}}
function.}

\item{type}{a character denoting which type of PCA plot should be created.
Default to "scatterplot". Type "biplot" shows eigenvectors.}

\item{group_by}{character; default to "sample_type", which makes a plot for
quality control. When set as "group", a PCA plot with respect to the groups
of samples with type 'Sample' is created.}

\item{types_to_display}{a vector of sample type names specifying which types
should be shown on the plot when type = "sample_type". Defaults to all.}

\item{threshold}{a value indicating the absolute minimum correlation between
a metabolite and any of the first two components, required for this
metabolite to be included on the biplot.}

\item{interactive}{logical. If \code{TRUE} (default), a ggiraph interactive
plot is returned; otherwise, a standard ggplot object is produced.}
}
\description{
\code{create_PCA_plot()} returns the scatterplot or biplot, based on the PCA.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_PCA_plot(dat)
create_PCA_plot(dat, type = "biplot", threshold = 0.3)
dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
create_PCA_plot(dat, group_by = "group")
create_PCA_plot(dat, type = "biplot", group_by = "group", threshold = 0.3)

}
