% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_PCA_plot}
\alias{create_PCA_plot}
\title{PCA plot}
\usage{
create_PCA_plot(
  dat,
  type = "sample_type",
  group_by = "sample_type",
  types_to_display = "all",
  threshold = NULL,
  interactive = TRUE
)
}
\arguments{
\item{type}{A character denoting which type of PCA plot should be created.
Default to "scatter". Type "biplot" shows eigenvectors.}

\item{group_by}{Character; default to "sample_type", which makes a plot for
quality control. When set as "group", a PCA plot with respect to the groups
of samples with type 'Sample' is created.}

\item{types_to_display}{A vector of sample type names specifying which types
should be shown on the plot when type = "sample_type". Defaults to all.}

\item{threshold}{A value indicating the minimum correlation between
a variable and any component, required for this  variable to be included
on the PCA biplot.}
}
\description{
PCA plot
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_PCA_plot(dat)
create_PCA_plot(dat, type = "biplot", threshold = 0.3)
dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
create_PCA_plot(dat, group_by = "group")
create_PCA_plot(dat, type = "biplot", group_by = "group", threshold = 0.3)

}
