% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_plot_of_2_metabolites}
\alias{create_plot_of_2_metabolites}
\title{Plot of two metabolites}
\usage{
create_plot_of_2_metabolites(dat, metabolite1, metabolite2)
}
\description{
Plot of two metabolites
}
