% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_qqplot}
\alias{create_qqplot}
\title{QQ plots of individual metabolite values before and after imputation}
\usage{
create_qqplot(dat, metabolite)
}
\arguments{
\item{dat}{a \code{raw_data} object, the output of the \code{\link[=read_data]{read_data()}} function.
The data have to be completed, for example using the \code{\link[=complete_data]{complete_data()}}
function.}

\item{metabolite}{a name of the metabolite of interest.}
}
\description{
\code{create_qqplot()} returns QQ plots for the specified metabolite
before and after imputation.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_qqplot(dat, "C0")

}
