% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{create_qqplot}
\alias{create_qqplot}
\title{Qqplots of individual metabolite values before and after imputation}
\usage{
create_qqplot(dat, metabolite)
}
\arguments{
\item{metabolite}{The name of metabolite of interest.}
}
\description{
Qqplots of individual metabolite values before and after imputation
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- complete_data(dat, "limit", "limit", "limit")
create_qqplot(dat, "C0")

}
