% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LOD_to_remove.R
\name{get_LOD_to_remove}
\alias{get_LOD_to_remove}
\title{Get metabolites to remove}
\usage{
get_LOD_to_remove(NA_info, treshold)
}
\arguments{
\item{NA_info}{Attribute of the raw_data object.}

\item{treshold}{Percentage value.}
}
\description{
Returns metabolite names having more NA values in each group than the given treshold.
}
