% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{get_LOD_to_remove}
\alias{get_LOD_to_remove}
\title{Get metabolites to remove}
\usage{
get_LOD_to_remove(NA_info, treshold)
}
\arguments{
\item{NA_info}{Attribute of the raw_data object.}

\item{treshold}{Percentage value.}
}
\description{
Returns metabolite names having more NA values in each group than the given treshold.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
get_LOD_to_remove(attr(dat, "NA_info"), 0.1)

}
