% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_model_summary}
\alias{get_model_summary}
\title{Get model summary}
\usage{
get_model_summary(model)
}
\arguments{
\item{model}{a model object, the output of \code{\link[=build_model]{build_model()}}.}
}
\description{
\code{get_model_summary()} returns a list containing: train dataset, test dataset
with predicted values, model coefficients, AUC value and ROC plot, for
a model.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
model <- build_model(dat, "group", "2", "Lasso")
get_model_summary(model)

}
