% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_model_summary}
\alias{get_model_summary}
\title{Get model summary}
\usage{
get_model_summary(model)
}
\description{
\code{get_models_info()} returns the summaries for models returned by
\code{\link[=build_model]{build_model()}}.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
model <- build_model(dat, "group", "2")
get_model_summary(model)

}
