% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{get_models_info}
\alias{get_models_info}
\title{Get information about models}
\usage{
get_models_info(models)
}
\arguments{
\item{response}{a string specifying the response variable.}

\item{level}{a string specifying which level of the response variable to
consider. Required only if response variable has over two levels.}
}
\description{
\code{build_models()} returns the linear models with and without SLOPE coś tam
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
models <- build_models(dat, "group", "2")
get_models_info(models)

}
