% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Heatmap of missing metabolites values}
\usage{
plot_heatmap(
  dat,
  plate_bar_code = NULL,
  include_title = FALSE,
  show_colors = TRUE
)
}
\arguments{
\item{dat}{A \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}} function}

\item{plate_bar_code}{A single plate bar code for selecting observations to
include in the plot. If NULL (default), a grid of plots is returned, where
each plot corresponds to a different plate bar code.}

\item{include_title}{Logical. Indicates whether the title with the
plate bar code should be included (only if \code{plate_bar_code} is not NULL).
Defaults to \code{FALSE}.}

\item{show_colors}{Logical. If \code{TRUE}, applies distinct colors to different
types of missing values, following the conventions used in Biocrates® files.}
}
\description{
Heatmap of missing metabolites values
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
plot_heatmap(dat, "1036372116-1 | 1036372121-1")

}
