% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_NA_percent}
\alias{plot_NA_percent}
\title{Barplot of missing metabolites values}
\usage{
plot_NA_percent(dat, type = NULL)
}
\arguments{
\item{type}{NULL, "Type" or "group"}
}
\description{
Barplot of missing metabolites values
}
