% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_NA_percent}
\alias{plot_NA_percent}
\title{Barplot of missing metabolites percents}
\usage{
plot_NA_percent(dat, type = "joint")
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}} function}

\item{type}{a character denoting which type of plot should be made. This
function accepts either "joint", "NA_type" or "group". TODO: what does it mean?}
}
\description{
Barplot of missing metabolites percents
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
plot_NA_percent(dat)
plot_NA_percent(dat, "NA_type")
dat <- add_group(dat, "group")
plot_NA_percent(dat, "group")

}
