% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_NA_percent}
\alias{plot_NA_percent}
\title{Barplot of missing metabolites values}
\usage{
plot_NA_percent(dat, type = NULL)
}
\arguments{
\item{dat}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}} function}

\item{type}{NULL, "Type" or "group"}
}
\description{
Barplot of missing metabolites values
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
plot_NA_percent(dat)
plot_NA_percent(dat, "NA_type")
dat <- add_group(dat, "group")
plot_NA_percent(dat, "group")
}
