% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{plot_NA_percent}
\alias{plot_NA_percent}
\title{Barplot of missing values percents}
\usage{
plot_NA_percent(dat, type = "joint", interactive = TRUE)
}
\arguments{
\item{dat}{A \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}} function}

\item{type}{A character denoting which type of plot should be made. This
function accepts either "joint", "NA_type" or "group".
Default type is "joint", which creates plot of missing values percents
in each metabolite. Types "NA_type" and "group" add the division into
all missing values types and levels in grouping column respectively.}
}
\description{
Barplot of missing values percents
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
print(plot_NA_percent(dat))
print(plot_NA_percent(dat, "NA_type"))
dat <- add_group(dat, "group")
print(plot_NA_percent(dat, "group"))

}
