% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{predict_probability}
\alias{predict_probability}
\title{Predict probabilities}
\usage{
predict_probability(model, new_dat = NULL)
}
\arguments{
\item{model}{a model object, the output of \code{\link[=build_model]{build_model()}}.}

\item{new_dat}{a cleaned metabolomics or compounds matrix. If no dataset is
provided, predictions for the test dataset are returned.}
}
\description{
\code{predict_probability()} returns the provided dataset with predicted
probabilities for the response variable.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
model <- build_model(dat, "group", "2", model = "Lasso")
predict_probability(model)

}
