% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{predict_probability}
\alias{predict_probability}
\title{Predict}
\usage{
predict_probability(model, new_dat = NULL)
}
\description{
Predict
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- add_group(dat, "group")
dat <- complete_data(dat, "limit", "limit", "limit")
model <- build_model(dat, "group", "2")
predict_probability(model)

}
