% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{remove_metabolites}
\alias{remove_metabolites}
\title{Adding metabolites to the attribute removed}
\usage{
remove_metabolites(raw_data, metabolites_to_remove, type)
}
\arguments{
\item{raw_data}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}}
function.}

\item{metabolites_to_remove}{metabolites to remove}

\item{type}{type of metabolites to remove}
}
\description{
Adding metabolites to the attribute removed
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
test_dat <- read_data(path)
attr(remove_metabolites(test_dat, "C0", "LOD"), "removed")

}
