% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{scale_color_metabocrates_continuous}
\alias{scale_color_metabocrates_continuous}
\title{MetaboCrates scale color continuous}
\usage{
scale_color_metabocrates_continuous()
}
\description{
MetaboCrates scale color continuous
}
\keyword{internal}
