% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funs.R
\name{scale_color_metabocrates_discrete}
\alias{scale_color_metabocrates_discrete}
\title{MetaboCrates scale color}
\usage{
scale_color_metabocrates_discrete(n = 12, ...)
}
\description{
MetaboCrates scale color
}
\keyword{internal}
