% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{show_data}
\alias{show_data}
\title{Show metabolites without removed}
\usage{
show_data(dat)
}
\arguments{
\item{dat}{A raw_data object.}
}
\description{
Returns metabolites without those in removed attribute.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- remove_metabolites(dat, "C0", type = "LOD")
show_data(dat)

}
