% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{show_ratios}
\alias{show_ratios}
\title{Show LOD ratios without removed metabolites}
\usage{
show_ratios(data)
}
\arguments{
\item{raw_data}{A raw_data object.}
}
\description{
Returns LOD ratios without metabolites in removed attribute.
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
dat <- read_data(path)
dat <- remove_LOD(dat, "C0")
show_ratios(dat)

}
