% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{unremove_all}
\alias{unremove_all}
\title{Setting attribute removed to NULL}
\usage{
unremove_all(raw_data, type)
}
\arguments{
\item{raw_data}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}}
function.}

\item{type}{type of metabolites to remove}
}
\description{
Setting attribute removed to NULL
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
test_dat <- read_data(path)
test_dat <- remove_metabolites(test_dat, "C0", "LOD")
attr(test_dat, "removed")
attr(unremove_all(test_dat, "LOD"), "removed")

}
