% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_funs.R
\name{unremove_metabolites}
\alias{unremove_metabolites}
\title{Removing given metabolites from the attribute removed}
\usage{
unremove_metabolites(raw_data, metabolites)
}
\arguments{
\item{raw_data}{a \code{\link{raw_data}} object. Output of \code{\link[=read_data]{read_data()}}
function.}

\item{metabolites}{a vector of metabolites to unremove}
}
\description{
Removing given metabolites from the attribute removed
}
\examples{
path <- get_example_data("small_biocrates_example.xls")
test_dat <- read_data(path)
test_dat <- remove_metabolites(test_dat, c("C0", "C1", "C2"), "LOD")
test_dat <- remove_metabolites(test_dat, "C0", "QC")
attr(test_dat, "removed")
attr(unremove_metabolites(test_dat, c("C0", "C1")), "removed")

}
